api/digitrecognizer/requirements.txt
numpy
pandas
math 
h5py
matplotlib.pyplot
scipy
tensorflow
os
random
cv2
sklearn
